api/requirements.txt
PyJWT==2.8.0
supabase==2.0.0